+++ packages/datadog_tracking_http_client/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-
+* Add the ability to ignore tracking on specific url patterns with `ignoreUrlPatterns` when using the attach configuration.
 
 ## 2.2.0